+++ client/Packages/com.beamable/Editor/UI/Toolbox/Components/ToolboxActionBarVisualElement/ToolboxActionBarVisualElement.cs
@@ -48,7 +48,7 @@ namespace Beamable.Editor.Toolbox.Components
 
 		public ToolboxActionBarVisualElement() : base(nameof(ToolboxActionBarVisualElement)) { }
 
-		public ToolboxModel Model { get; set; }
+		private IToolboxViewService Model { get; set; }
 
 		private Button _categoryButton;
 		private Button _typeButton;
@@ -61,6 +61,8 @@ namespace Beamable.Editor.Toolbox.Components
 		public override void Refresh()
 		{
 			base.Refresh();
+
+			Model = provider.GetService<IToolboxViewService>();
 
 			var contentButton = Root.Q<Button>("contentManager");
 			contentButton.clickable.clicked += async () => { await ContentManagerWindow.Init(); };
