+++ client/Packages/com.beamable/Editor/UI/Toolbox/Models/ToolboxViewService.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6be6bb1f5106f44c685a942d1f1853f8
+guid: cce0f50ff5e07604eb8f00d0dbbea8c1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
