+++ client/Packages/com.beamable/Editor/UI/Toolbox/ToolboxWindow.cs
@@ -5,6 +5,9 @@ using Beamable.Editor.NoUser;
 using Beamable.Editor.Toolbox.Components;
 using Beamable.Editor.Toolbox.Models;
 using Beamable.Editor.UI;
+
+using Beamable.Common.Dependencies;
+
 using System;
 using UnityEditor;
 using UnityEditor.VspAttribution.Beamable;
@@ -24,7 +27,6 @@ namespace Beamable.Editor.Toolbox.UI
 {
 	public class ToolboxWindow : BeamEditorWindow<ToolboxWindow>
 	{
-
 		static ToolboxWindow()
 		{
 			WindowDefaultConfig = new BeamEditorWindowInitConfig()
@@ -53,7 +55,7 @@ namespace Beamable.Editor.Toolbox.UI
 
 		private ToolboxContentListVisualElement _contentListVisualElement;
 
-		private ToolboxModel _model;
+		private IToolboxViewService _model;
 		private ToolboxAnnouncementListVisualElement _announcementListVisualElement;
 
 		protected override void Build()
@@ -68,10 +70,11 @@ namespace Beamable.Editor.Toolbox.UI
 			// Refresh if/when the user logs-in or logs-out while this window is open
 			ActiveContext.OnUserChange += _ => BuildWithContext();
 
+			_model = ActiveContext.ServiceScope.GetService<IToolboxViewService>();
+
 			// Force refresh to build the initial window
 			_model?.Destroy();
 
-			_model = new ToolboxModel();
 			_model.UseDefaultWidgetSource();
 			_model.Initialize();
 
@@ -116,18 +119,15 @@ namespace Beamable.Editor.Toolbox.UI
 			root.Add(_windowRoot);
 
 			_actionBarVisualElement = root.Q<ToolboxActionBarVisualElement>("actionBarVisualElement");
-			_actionBarVisualElement.Model = _model;
 			_actionBarVisualElement.Refresh();
 
 			_breadcrumbsVisualElement = root.Q<ToolboxBreadcrumbsVisualElement>("breadcrumbsVisualElement");
 			_breadcrumbsVisualElement.Refresh();
 
 			_contentListVisualElement = root.Q<ToolboxContentListVisualElement>("contentListVisualElement");
-			_contentListVisualElement.Model = _model;
 			_contentListVisualElement.Refresh();
 
 			_announcementListVisualElement = root.Q<ToolboxAnnouncementListVisualElement>();
-			_announcementListVisualElement.Model = _model;
 			_announcementListVisualElement.Refresh();
 			_announcementListVisualElement.OnHeightChanged += AnnouncementList_OnHeightChanged;
